ECO-3042:Braintree/Adyen Credit card modules overwrite each other
