+++ src/SprykerEco/Yves/Braintree/Theme/default/components/molecules/braintree-pay-pal-express/braintree-pay-pal-express.twig
@@ -1,8 +1,8 @@
 {% extends model('component') %}
 
 {% define config = {
-    name: 'pay-pal-express',
-    tag: 'pay-pal-express'
+    name: 'braintree-pay-pal-express',
+    tag: 'braintree-pay-pal-express'
 } %}
 
 {% define attributes = {
