+++ src/SprykerEco/Yves/Braintree/Theme/default/components/molecules/braintree-pay-pal/braintree-pay-pal.twig
@@ -1,8 +1,8 @@
-{% extends molecule('payment-form', 'Braintree') %}
+{% extends molecule('braintree-payment-form', 'Braintree') %}
 
 {% define config = {
-    name: 'pay-pal',
-    tag: 'pay-pal'
+    name: 'braintree-pay-pal',
+    tag: 'braintree-pay-pal'
 } %}
 
 {% define attributes = {
