+++ src/SprykerEco/Yves/Braintree/Theme/default/components/molecules/braintree-pay-pal/index.ts
@@ -1,0 +1,3 @@
+import register from 'ShopUi/app/registry';
+export default register('braintree-pay-pal', () => import(/* webpackMode: "lazy" */'./braintree-pay-pal'));
+
