+++ src/SprykerEco/Yves/Braintree/Theme/default/components/molecules/braintree-payment-form/braintree-payment-form.twig
@@ -1,8 +1,8 @@
 {% extends model('component') %}
 
 {% define config = {
-    name: 'payment-form',
-    tag: 'payment-form'
+    name: 'braintree-payment-form',
+    tag: 'braintree-payment-form'
 } %}
 
 {% define data = {
