--- src/SprykerEco/Yves/Braintree/Theme/default/components/molecules/pay-pal/index.ts
@@ -1,3 +1,0 @@
-import register from 'ShopUi/app/registry';
-export default register('pay-pal', () => import(/* webpackMode: "lazy" */'./pay-pal'));
-
